Feat/usercontext

+++ lms/src/components/sideBar/SideBar.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useContext} from "react";
 import { IoDocumentOutline, IoSettingsOutline } from "react-icons/io5";
 import { MdOutlineAssignment } from "react-icons/md";
 import { FaBook } from "react-icons/fa";
@@ -6,19 +6,43 @@ import { VscHome } from "react-icons/vsc";
 import { FiLogIn } from "react-icons/fi";
 import styles from "./SideBar.module.css"
 import Logo from "../images/LogoWhite.png"; 
-import { NavLink, Link, useNavigate } from "react-router-dom";
+import { NavLink, Link } from "react-router-dom";
 import { AiOutlinePieChart } from 'react-icons/ai';
+import axios from "axios";
+import { UserContext } from '../../utils/UserContext';
 
 
 
 function SideBar(){
-    const navigate = useNavigate();
- 
-    const handleLogout = () => {
-        localStorage.clear();
 
-        navigate('/')
-    }
+    const user = useContext(UserContext)
+ 
+    const handleLogout = async () => {
+    
+        let token = localStorage.getItem("token")
+        // let email = localStorage.getItem("userEmail")
+        const email = user.userData.data.Email
+          const data = {
+            email: email,
+          }
+          const url = "https://kidtots.onrender.com/student/logout"
+        
+          await axios.post(url, data, {
+            headers:{
+            "Authorization": "Bearer" + token,
+            "Content-Type": "application/json",
+            }
+           }).then((response) =>{
+         
+            if(response.status === 200){
+                localStorage.clear();
+            // setEmail("");
+            }
+           }).then((error) =>{
+
+           })     
+        }   
+  
 
 return(
     <>
@@ -89,7 +113,7 @@ return(
                 <div className={styles.sideBarFooter}>
                     <div>
                         <Link to="/">
-                        <div>
+                        <div onClick={handleLogout}>
                         <i><FiLogIn /></i><p>Logout</p>
                         </div>
                         </Link>
